Rewrite d.ts in external module definition format
