+++ types/test/tsconfig.json
@@ -7,6 +7,8 @@
   "files": [
     "index.ts",
     "../index.d.ts",
+    "../../logger.d.ts",
+    "../vue.d.ts",
     "../typings/index.d.ts"
   ]
 }
